Split up script_edit_dialog.py and add tests for SequenceWidget et al.
No functional change intended.
